+++ client/src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import { Chat } from './chat/chat.js';
-import { ChatList } from './chat/chat-list.js';
+import { Chat } from './pages/chat-room/chat.js';
+import { ChatList } from './pages/chat-list/chat-list.js';
 
 function App() {
   return (
