+++ client/src/index.js
@@ -1,6 +1,5 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import "./chat/chat.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
